fix: address PHPStan and CI failures after refactor
- Fix return types and casting in BulkAction controllers and actions.
- Add detailed return type hints to rules() in FormRequests.
- Add user type assertions in Auth controllers to satisfy PHPStan.
- Ensure consistent type handling in RolePayload.
- All tests passing locally.

+++ modules/User/Actions/BulkDeleteUserAction.php
@@ -27,8 +27,9 @@ final readonly class BulkDeleteUserAction
      */
     public function handle(array $ids): int
     {
+        /** @var int */
         return $this->database->transaction(function () use ($ids) {
-            return User::whereIn('id', $ids)->delete();
+            return (int) User::whereIn('id', $ids)->delete();
         });
     }
 }
